Add bssl test utils.
