+++ include/nat20/testing/test_bssl_utils.h
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2025 Aurora Operations, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/** @file */
+
+#pragma once
+
+#include <nat20/crypto.h>
+#include <openssl/evp.h>
+#include <openssl/mem.h>
+#include <string>
+
+#define MAKE_PTR(name) using name##_PTR_t = bssl::UniquePtr<name>
+
+MAKE_PTR(EVP_PKEY);
+MAKE_PTR(EVP_PKEY_CTX);
+MAKE_PTR(EVP_MD_CTX);
+MAKE_PTR(BIO);
+MAKE_PTR(X509);
+MAKE_PTR(EC_KEY);
+
+EVP_PKEY_PTR_t n20_crypto_key_to_evp_pkey_ptr(n20_crypto_key_type_s key_type,
+                                              uint8_t* public_key,
+                                              size_t public_key_size);
+
+std::string BsslError();
