+++ src/testing/test_bssl_utils.cpp
@@ -1,0 +1,73 @@
+/*
+ * Copyright 2025 Aurora Operations, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include <gtest/gtest.h>
+#include <nat20/testing/test_bssl_utils.h>
+#include <openssl/ec_key.h>
+#include <openssl/err.h>
+
+EVP_PKEY_PTR_t n20_crypto_key_to_evp_pkey_ptr(n20_crypto_key_type_s key_type,
+                                              uint8_t* public_key,
+                                              size_t public_key_size) {
+    switch (key_type) {
+        case n20_crypto_key_type_ed25519_e: {
+            auto key = EVP_PKEY_PTR_t(
+                EVP_PKEY_new_raw_public_key(EVP_PKEY_ED25519, NULL, public_key, public_key_size));
+            if (!key) {
+                ADD_FAILURE();
+                return nullptr;
+            }
+
+            return key;
+        }
+        case n20_crypto_key_type_secp256r1_e:
+        case n20_crypto_key_type_secp384r1_e: {
+            auto ec_key = EC_KEY_PTR_t(EC_KEY_new_by_curve_name(
+                key_type == n20_crypto_key_type_secp256r1_e ? NID_X9_62_prime256v1
+                                                            : NID_secp384r1));
+            if (!ec_key) {
+                ADD_FAILURE();
+                return nullptr;
+            }
+
+            auto ec_key_p = ec_key.get();
+            uint8_t const* p = public_key;
+            if (!o2i_ECPublicKey(&ec_key_p, &p, public_key_size)) {
+                ADD_FAILURE();
+                return nullptr;
+            }
+
+            auto key = EVP_PKEY_PTR_t(EVP_PKEY_new());
+            if (!key || !EVP_PKEY_assign_EC_KEY(key.get(), ec_key.release())) {
+                ADD_FAILURE();
+                return nullptr;
+            }
+
+            return key;
+        }
+        default: {
+            ADD_FAILURE();
+            return nullptr;
+        }
+    }
+}
+
+std::string BsslError() {
+    char buffer[2000];
+    auto error = ERR_get_error();
+    ERR_error_string_n(error, buffer, 2000);
+    return std::string(buffer);
+}
